Merge: Road Classes refacted step 2/2 and Merge w/branch GameMenu

+++ src/main/entity/Enemy.java
@@ -41,7 +41,7 @@ public class Enemy extends Entity {
         //g2.drawImage(bufferedImage, x, 570, 240, 240, null);
     }
 
-    public int getSpeed() {
+    public double getSpeed() {
         return speed;
     }
